🚚 move sponsor component
