+++ frontend/src/pages/ProjectDetails/Overview/OverviewPanel/Sponsor.tsx
